add small-test.png file and modify module path error
