+++ src/server/aws-s3.js
@@ -30,14 +30,15 @@ class MockS3 {
   deleteObject(params = {}, cb = () => {}) {
     cb(null, params);
   }
-}*/
-export const S3Utils ={
-   imgToBase64: (imgPath, cb) => {
+}
+*/
+export const S3Utils = {
+  imgToBase64: (imgPath, cb) => {
     fs.readFile(imgPath, (err, image) => {
       return cb(err, new Buffer(image, 'binary').toString('base64'));
     });
   }
-}
+};
 export class S3Connector {
   constructor() {
     //if (process.env.NODE_ENV === 'test') {
@@ -47,7 +48,7 @@ export class S3Connector {
       // You have to change value of awsConfig of 'config/test.json'
       //   as 'aws.json' and fill 'aws.json' with proper values for real s3 instance test
       // this code is just for No more request at CI time (aws pretty expensive)
-     // this.s3instance = new MockS3(config.awsConfig);
+    //  this.s3instance = new MockS3(config.awsConfig);
     //} else {
       this.s3instance = new AWS.S3();
     //}
