+++ src/server/key-utils.js
@@ -2,7 +2,6 @@ import geohash from 'ngeohash';
 import uuid from 'uuid4';
 
 export const MAX_TIME = 10000000000000;
-
 export const DEFAULT_PRECISON = 8;
 export const STATE = {ALIVE: '0', EXPIRED: '1', REMOVED: '2'};
 export const ENTITY = {
@@ -15,12 +14,18 @@ export const ENTITY = {
   SAVED_POST: 'saved-post',
   REACT_POST: 'react-post'
 };
+export const CATEGORY = {
+  EVENT: 'event',
+  WARNING: 'warning',
+  FACILITY: 'facility'
+};
 export const STATE_CODE_POS = 5;
 export const GEOHASH_START_POS = 5;
 export const GEOHASH_END_POS = GEOHASH_START_POS + DEFAULT_PRECISON - 1;
 export const UUID_START_POS = 14;
 export const REF_GEOHASH_START_POS = 7;
-
+const DELIMITER_NUM_IN_KEY = 6;
+const TIMEHASH_LENGTH = 50;
 export class Timestamp {
   constructor(date) {
     this.timeStamp = MAX_TIME - Number(new Date(date));
@@ -47,9 +52,13 @@ export class KeyMaker {
     return this.uuid;
   }
 }
-
-
 export const KeyUtils = {
+  isOriginKey: (key) => {
+    return ((key.match(/-/g) || []).length === DELIMITER_NUM_IN_KEY) ? true : false;
+  },
+  getTimeHash: (key) => {
+    return key.substring(key.length - TIMEHASH_LENGTH, key.length);
+  },
   getReversedTime: (time = new Date()) => {
     return MAX_TIME - Number(new Date(time));
   },
@@ -66,6 +75,13 @@ export const KeyUtils = {
       key = `${key}-${option}`;
     });
     return `${key}-${timeHash}`;
+  },
+  getIdxKeys: (lat, lng, timeHash) => {
+    const keys = [];
+    for (let i = 1; i <= DEFAULT_PRECISON; i = i + 1) {
+      keys.push(`item-${STATE.ALIVE}-${geohash.encode(lat, lng, i)}-${timeHash}`);
+    }
+    return keys;
   },
   getKeysByArea: (lat, lng, precision) => {
     const centerGeohash = geohash.encode(lat, lng, precision);
@@ -95,7 +111,3 @@ export const KeyUtils = {
      db.put(idxKey, ~~~);
   4. return result;
 */
-
-
-
-
