+++ src/test/node/index.js
@@ -4,7 +4,7 @@ require('source-map-support').install();
 // Add more tests by importing them here.
 
 import './server';
-import './items/key-utils';
+import './key-utils';
 import './items/validator';
 import './items/controllers';
 import './database';
