+++ src/test/node/key-utils.js
@@ -2,7 +2,7 @@ import test from 'tape';
 import geohash from 'ngeohash';
 import {KeyMaker, KeyUtils, STATE, DEFAULT_PRECISON, MAX_TIME, Timestamp,
         GEOHASH_START_POS, GEOHASH_END_POS, REF_GEOHASH_START_POS, STATE_CODE_POS}
-        from '../../../server/key-utils';
+        from '../../server/key-utils';
 
 const mockData = {
   lat: 37.565398,
